requirements.txt
Mako~=1.1
PyYAML~=6.0
absl-py~=2.1
docker==7.1.0
google-api-python-client~=1.12
google-cloud-secret-manager~=2.1
google-cloud-monitoring~=2.18
grpcio~=1.60
grpcio-health-checking~=1.60
grpcio-tools~=1.60
grpcio-channelz~=1.60
kubernetes~=27.2
six~=1.13
tenacity~=6.2
typing-extensions~=4.10
packaging~=23.1
Pygments~=2.9
python-dateutil~=2.8
protobuf~=4.24
requests~=2.31.0
xds-protos==1.58.0rc1
